feature: add cart - api

+++ src/main/java/vn/edu/hcmute/foodapp/dto/request/AddCartItemRequest.java
@@ -16,5 +16,5 @@ public class AddCartItemRequest {
     @Min(value = 1, message = "quantity must be greater than 0")
     Integer quantity;
 
-    List<Long> selectedMenuItemOptionIds;
+    List<Integer> selectedMenuItemOptionIds;
 }
